cleanup types (checkpoint #2)
